Bug fixes: * oct_geom_opt is no longer existent. CRAY compiler compiled because it was declared, although not used. * Default integer type should be used everywhere to avoid problems...

+++ src/liboct.F90
@@ -130,7 +130,7 @@ module liboct
 
     subroutine oct_wfs_list(str, l)
       character(len=*), intent(in) :: str
-      integer(4), intent(out) :: l(32)
+      integer, intent(out) :: l(32)
     end subroutine oct_wfs_list
 
     real(8) function oct_ran_gaussian(r, sigma)
